requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
blinker==1.9.0
brotli==1.2.0
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cssselect2==0.9.0
Deprecated==1.2.18
fastapi==0.115.8
Flask==3.1.3
fonttools==4.62.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.2.0
lxml==5.3.1
MarkupSafe==3.0.3
packaging==25.0
pdfkit==1.0.0
pillow==12.2.0
pycparser==3.0
pydantic==2.11.4
pydantic_core==2.33.2
pydyf==0.12.1
PyMySQL==1.1.2
pyphen==0.17.2
python-dotenv==1.2.2
requests==2.32.3
selectolax==0.3.27
slowapi==0.1.9
sniffio==1.3.1
starlette==0.45.3
tinycss2==1.5.1
tinyhtml5==2.1.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.0
webencodings==0.5.1
Werkzeug==3.1.8
wrapt==1.17.2
zopfli==0.4.1
